chore(tracing): move snapshotterInjected and mergeTraceFiles into packages/tracing/writer
snapshotterInjected.ts is pure DOM logic (no Playwright internals)
and now lives next to the rest of the writer; snapshotter.ts
imports it via @tracing/writer/snapshotterInjected.

mergeTraceFiles, previously a private helper at the bottom of
testTracing.ts, is now a generic library function. The
test-trace-entry-name special case is exposed as an
optional `keepEntryName` parameter; the test runner passes
`testTraceEntryName`.
