+++ packages/playwright-core/src/server/trace/recorder/DEPS.list
@@ -1,5 +1,6 @@
 [*]
 @isomorphic/**
+@tracing/**
 @utils/**
 ../../
 ../../har/
