+++ packages/tracing/src/writer/snapshotterInjected.ts
